requirements.txt
alembic==1.11.1
aniso8601==6.0.0
astroid==2.2.5
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
Click==7.0
ecdsa==0.13.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.5.0
Flask-WTF==1.1.1
future==0.17.1
greenlet==2.0.2
gunicorn==20.0.4
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==4.3.18
itsdangerous==2.0.1
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.2.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==23.1
pluggy==0.13.1
psycopg2==2.9.6
psycopg2-binary==2.8.2
py==1.11.0
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pytest==6.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose-cryptodome==1.3.2
pytz==2019.1
ruamel.yaml==0.16.5
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.2
typed-ast==1.4.2
typing-extensions==4.6.3
Werkzeug==2.0.3
wrapt==1.11.1
WTForms==3.0.1
zipp==3.15.0
